isAdmin

+++ api/usersRouter.js
@@ -128,13 +128,15 @@ usersRouter.get("/", async (req, res, next) => {
 */
 
 usersRouter.delete("/:userId", requireUser, async (req, res, next) => {
-  try {
-    const { userId } = req.params;
-    const deleteSingleUser = await deleteUser(userId);
-    res.send(deleteSingleUser);
-  } catch (error) {
-    console.log("error deleting user");
-    throw error;
+  if ((req.user.isAdmin = true)) {
+    try {
+      const { userId } = req.params;
+      const deleteSingleUser = await deleteUser(userId);
+      res.send(deleteSingleUser);
+    } catch (error) {
+      console.log("error deleting user");
+      throw error;
+    }
   }
 });
